NKARDemo : cible Android — l'AR passe sur telephone
Etage P1 du plan de perception. Le portage est plus mince qu'il n'y paraissait :
la chaine AR — marqueurs, session, monde, suivi — est du C++ pur sans la
moindre dependance plateforme, elle traverse telle quelle. Restaient la camera,
dont le backend Camera2 NDK etait deja ecrit, et l'empaquetage.

Un seul defaut de compilation, et c'est un piege deja connu du depot :
NkChrono::Sleep expose une surcharge int64 et une float64, si bien qu'un
litteral nu est ambigu. Le backend camera Android n'avait jamais ete compile —
c'est la premiere fois qu'il passe le compilateur.

L'APK sort signe avec les quatre architectures. Les quatre restent declarees
dans le descripteur, qui dit ce que l'application SUPPORTE ; la restriction a
une seule architecture pendant la mise au point se fait en ligne de commande,
ou est sa place.

Dette relevee au passage, a la charge de NKCamera : NkAndroidCameraBackend.h
utilise seize fois la bibliotheque standard (std::mutex, std::unique_lock),
ce qui contrevient a la regle zero-STL du depot.

+++ Kernel/Runtime/NKCamera/src/NKCamera/Backend/NkAndroidCameraBackend.h
@@ -381,7 +381,7 @@ namespace nkentseu {
 					const NkElapsedTime start = NkChrono::Now();
 					while (!mHasFrame && (NkChrono::Now() - start).seconds < 3.0) {
 						lk.unlock();
-						NkChrono::Sleep(10);
+						NkChrono::Sleep(int64(10)); // surcharges int64/float64 : un litteral nu est ambigu
 						lk.lock();
 					}
 				}
